pyproject.toml
[project]
name = "documentation-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "crawl4ai>=0.7.8",
    "fastapi>=0.128.0",
    "llama-index>=0.14.12",
    "llama-index-embeddings-ollama>=0.8.6",
    "llama-index-llms-ollama>=0.9.1",
    "llama-index-vector-stores-chroma>=0.5.5",
    "llama-index-vector-stores-pinecone>=0.7.1",
    "pinecone>=7.3.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "uvicorn>=0.40.0",
]
